refact struct
